Separated game into own file
- Created `game.jsx` file and copy/pasted game code from `index.js`
- Added `import Game from './game.jsx';` at top of new file
- Wrapped `<Game />` in div to `ReactDOM.render`
- Added `import React from 'react';` to top of `game.jsx`
