+++ public/index.html
@@ -19,7 +19,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>My React App</title>
+    <title>React App: Tic-Tac-Toe</title>
   </head>
   <body>
     <noscript>
