+++ src/index.css
@@ -1,6 +1,7 @@
 body {
   font: 14px "Century Gothic", Futura, sans-serif;
   margin: 20px;
+  background-color: #ccc;
 }
 
 ol, ul {
